chore: refactor dkg tests
