+++ src/eddsa/mod.rs
@@ -1,5 +1,4 @@
 //! This module serves as a wrapper for Ed25519 scheme.
-pub mod dkg_ed25519;
 pub mod sign;
 #[cfg(test)]
 mod test;
